requirements_complete.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
Pillow==10.1.0
openai==1.3.5
sqlalchemy==2.0.23
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
scikit-image==0.22.0
beautifulsoup4==4.12.2
pandas==2.1.4
aiofiles==23.2.1
google-generativeai==0.3.2
